requirements.txt
# Core LangChain and LangGraph dependencies
langgraph>=0.2.0
langchain>=0.2.0
langchain-openai>=0.1.0
langchain-anthropic>=0.1.0
langchain-core>=0.2.0

# Web framework and API
fastapi>=0.100.0
uvicorn[standard]>=0.23.0
jinja2>=3.1.0

# Data handling and validation
pydantic>=2.0.0
python-dotenv>=1.0.0
pymongo>=4.5.0

# Development and testing
pytest>=7.4.0
pytest-asyncio>=0.21.0
black>=23.0.0
flake8>=6.0.0

# Additional utilities
requests>=2.31.0
numpy<2
pandas>=2.0.0
typing-extensions>=4.7.0

# Fix for httpx compatibility
httpx>=0.27.0